% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dividends.R
\name{adjustOHLC_DividendWise_tibble_obj}
\alias{adjustOHLC_DividendWise_tibble_obj}
\title{adjustOHLC_DividendWise_tibble_obj}
\usage{
adjustOHLC_DividendWise_tibble_obj(tibble_obj, dividend_dates_tibble)
}
\arguments{
\item{dividend_dates_tibble}{}
}
\value{
original tibble objec timeseries data with OHLC now adjusted for all
  dividends.
}
\description{
Does the same thing as AdjustOHLC(), but it is our in house function for
tibble objects. Calculates the dividend adjustment ratio and returns adusted
OHLC time series. This function is part of a workflow, must
get_dividends_for_this_tibble() first.
}
