% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splits.R
\name{adjustOHLC_SplitWise_tibble_obj}
\alias{adjustOHLC_SplitWise_tibble_obj}
\title{adjustOHLC_Splitwise_tibble_obj}
\usage{
adjustOHLC_SplitWise_tibble_obj(tibble_obj, split_dates_tibble)
}
\arguments{
\item{split_dates_tibble}{}
}
\description{
Currently, not used because all of the Yahoo finance data is split adjusted.
}
