% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_yahoo_quantmod.R
\name{adjustOHLC_wrapper}
\alias{adjustOHLC_wrapper}
\title{adjustOHLC_wrapper}
\usage{
adjustOHLC_wrapper(tickers)
}
\arguments{
\item{tickers}{list of tickers, strings, for whose prices will be adjusted
based on the close-adjusted adjustment.}
}
\value{
nothing is returned, the existing xts objects are updated to reflect
  the adjusted prices, OHLC, plus a sanity check, ticker.AdjDiff to show us
  the difference between Yahoo's adjusted close and our calculated close with
  divident adjustment,.
}
\description{
AdjustOHLC in quantmod is NOT reliable. Yahoo data is currently all split
adjusted, but only the adjusted close is split AND dividend adjusted. Here we
are downloading the dividend data and doing the ratio calc ourselves. We are
downloading the divident data from Yahoo! as source.
}
