% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_av_quantmod.R
\name{av_Main_retrieve_and_write}
\alias{av_Main_retrieve_and_write}
\title{av_Main_retrieve_and_write}
\usage{
av_Main_retrieve_and_write(tickers, path_to_ticker_dir, remote_src = "av",
  start_date = "2017-11-01")
}
\arguments{
\item{tickers}{list of tickers}

\item{remote_src}{"av" for alphavantage}

\item{start_date}{start date of tick pull, doesn't get considered with
alphavantage, doing a full tick pull everytime right now, output.size =
"full".}
}
\value{
list of tickers
}
\description{
Alphavantage main retrieve and write function, retrieve data and write out to
its own directory.
}
