% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_quantmod.R
\name{calc_AdjCl_to_AdjCl_returns}
\alias{calc_AdjCl_to_AdjCl_returns}
\title{calc_AdjCl_to_AdjCl_returns}
\usage{
calc_AdjCl_to_AdjCl_returns(ticker)
}
\arguments{
\item{ticker}{the ticker string representing the xts time series data object}
}
\description{
calculates the returns of the adjusted close at t and the adjusted close at t-1. Uses the
Return.calculate functionality. Creates a new column named "AdAdDailyRets"
and appends it onto existing xts object and assigns this new object to memory
with name "ticker".
}
