% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_quantmod.R
\name{calc_change_in_volume}
\alias{calc_change_in_volume}
\title{calc_change_in_volume}
\usage{
calc_change_in_volume(ticker)
}
\arguments{
\item{ticker}{}
}
\value{
nothing
}
\description{
Calculates change in volume, row over row. Extends quantmod. Uses
PerformanceAnalytics Return.calculate. Requires a call to getSymbol, with
auto.assign set to TRUE.
}
