% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_quantmod.R
\name{calc_Close_to_Close_returns}
\alias{calc_Close_to_Close_returns}
\title{calc_Close_to_Close_returns}
\usage{
calc_Close_to_Close_returns(ticker)
}
\arguments{
\item{ticker}{}
}
\value{
returns nothing.
}
\description{
calculates the returns of the close at t and the close at t-1. Uses the
Return.calculate from PerformanceAnalytics functionality. Creates a new
column named "ClClDailyRets" and appends it onto existing xts object in
memory and assigns this new object to memory with name "ticker". Requires
that there be an object named ticker in memory, that contains a "Close"
column since we are also usind QuantMod here. This function is part of a
workflow. You must have called getSymbols() before this function with the
auto.assign option set to TRUE.
}
