% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_quantmod.R
\name{calc_Open_to_Open_returns}
\alias{calc_Open_to_Open_returns}
\title{calc_Open_to_Open_returns}
\usage{
calc_Open_to_Open_returns(ticker)
}
\arguments{
\item{ticker}{the ticker string representing the xts time series data object}
}
\value{
returns nothing
}
\description{
calculates the returns of the open at t and the open at t-1. Uses the
Return.calculate functionality. Creates a new column named "OpOpDailyRets"
and appends it onto existing xts object and assigns this new object to memory
with name "ticker".
}
