% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_quantmod.R
\name{calc_rolling_geom_mean_of_returns}
\alias{calc_rolling_geom_mean_of_returns}
\title{calc_rolling_geom_mean_of_returns}
\usage{
calc_rolling_geom_mean_of_returns(xts_obj, window = 21)
}
\arguments{
\item{window}{moving window size/length}
}
\value{
xts_obj with only the newly calculated rolling geometric mean.
}
\description{
Same as calc_rolling_mean_of_returns but returns the geometric mean instead of arithmetic mean.
}
