% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_quantmod.R
\name{calc_rolling_mean_of_returns}
\alias{calc_rolling_mean_of_returns}
\title{calc_rolling_mean_of_returns}
\usage{
calc_rolling_mean_of_returns(xts_obj, window = 21)
}
\arguments{
\item{window}{length of the window to caculate the mean over.}
}
\value{
an xts object containing the moving average of time series in xts_obj
}
\description{
Takes as input an xts time series object (of price or returns, not restricted
to returns), and caculates a moving window average, of length "window". This
function uses the rollapply functionality from xts/zoo.
}
