% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_quantmod.R
\name{calc_rolling_sd_of_returns}
\alias{calc_rolling_sd_of_returns}
\title{calc_rolling_std_of_returns}
\usage{
calc_rolling_sd_of_returns(xts_obj, window = 21)
}
\arguments{
\item{xts_obj}{xts object containing returns (daily, weekly, monthly, etc...)}

\item{window}{moving window size/length}
}
\value{
an xts object with the moving standard deviation of price data in
  input xts_obj. Column names in the returne object are not transformed, they
  are original.
}
\description{
calculates a moving window of standard deviation, of window size "window".
Returns a new xts object with only this column in it.
}
