% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_quantmod.R
\name{calc_rolling_sharpe_of_returns}
\alias{calc_rolling_sharpe_of_returns}
\title{calc_rolling_sharpe_of_returns}
\usage{
calc_rolling_sharpe_of_returns(xts_obj, window = 21,
  mean_type = "geometric", scale = 252)
}
\arguments{
\item{xts_obj}{xts input object of raw price data, a result of a call to getSymbols(), most simply}

\item{window}{default 21 (1 month of trading days)}

\item{mean_type}{default "geometric", optionally "arithmetic"}

\item{scale}{the number of period in 1 year; depends on the periodicity of the data in xts_obj}
}
\value{
xts with rolling sharpe of returns
#' @export
}
\description{
calc_rolling_sharpe_of_returns
}
