% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_yahoo_quantmod.R
\name{check_Rcompat_ticker_names}
\alias{check_Rcompat_ticker_names}
\title{check_Rcompat_ticker_names}
\usage{
check_Rcompat_ticker_names(ticker, path_to_ticker_dir,
  start_date = "2007-01-01")
}
\arguments{
\item{ticker}{ticker as a string}

\item{path_to_ticker_dir}{path to directory where to write csv file out.}

\item{start_date}{what date to start the tick pull from when the getSymbols()
quantmod func is called.}
}
\value{
the ticker as a string if it was modified, NA if it was not modified.
}
\description{
Takes a ticker string and removes any R incompatible text (reserved keywords
& symbols). Renames the ticker, pulls data on the ticker using getSymbols()
from quantmod. Calls adjustYahooOHLC() method to adjust the OHL prices in the
same manner the close price. Renames the columns prior to write out to csv so
that column names don't have any R incompatible characters. Calls the
write_ticker_csv function to write the ticker timeseries to file.
}
