% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_av_quantmod.R
\name{check_Rcompat_ticker_names_av}
\alias{check_Rcompat_ticker_names_av}
\title{check_Rcompat_ticker_names_av}
\usage{
check_Rcompat_ticker_names_av(path_to_ticker_dir, start_date, remote_src)
}
\arguments{
\item{path_to_ticker_dir}{ticker directory for writing out alphavantage ticks}

\item{remote_src}{"av" for alphavantage}
}
\value{
a list of tickers that were changed.
}
\description{
Check for R compatibiity of ticker names. Anything with a '-' or tickers in
list (C, T, F, NA) are changed as these are reserved characters/strings in R.
Also, calls getSymbols and write_ticker_csv for those tickers that are
changed. Finally, returns a list of those tickers that were changed.
}
