% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore_stats_funcs.R
\name{close_less_open_div_range}
\alias{close_less_open_div_range}
\title{close_less_open_div_range}
\usage{
close_less_open_div_range(v)
}
\arguments{
\item{v}{}
}
\value{
a "day sharpe" value based on open to close return and day's range.
}
\description{
Returns a syntheic "day sharpe" for lack of a better term. It is a helper
method used by get_close_less_open_div_range. Take the day's close price,
minus the day's open price, and divides it by the range (high - low).
}
