% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_quantmod.R
\name{create_balanced_weighted_portfolio}
\alias{create_balanced_weighted_portfolio}
\title{create_balanced_weighted_portfolio}
\usage{
create_balanced_weighted_portfolio(portfolio_returns, weights_vector,
  rebalance_on = "months")
}
\arguments{
\item{rebalance_on}{}
}
\description{
Incomplete function, not yet ready to be used.
GM - 2/16/2018
}
