% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_quantmod.R
\name{create_portfolio_xts}
\alias{create_portfolio_xts}
\title{create_portfolio_xts}
\usage{
create_portfolio_xts(tickers, price_type = "Ad")
}
\arguments{
\item{tickers}{a list of tickers to create the portfolio xts obect}

\item{price_type}{the price type : "Hi", "Lo", "Op", "Cl", "Ad". Only
price_type price for each ticker in tickers.}
}
\value{
a new xts object representing the portfolio of tickers with price type price.
}
\description{
creates a portfolio of xts tickers objects with only the given price type in
the mew merged xts object. The price type options are: "Hi", "Lo", "Op",
"Cl", and "Ad". Returned from this function is a new xts with multiple
tickers with the price_type as one colunn per ticker.
}
