% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_quantmod.R
\name{create_weekly_max_min_mean_std_of_Ad_price}
\alias{create_weekly_max_min_mean_std_of_Ad_price}
\title{create_weekly_max_min_mean_std_of_Ad_price}
\usage{
create_weekly_max_min_mean_std_of_Ad_price(xts_obj)
}
\arguments{
\item{xts_obj}{an xts object as would be returned by
getSymbol(<stock_ticker>). This function is very strict in its input and
output and requires there be a "Adjusted" (adjusted close) price column.}
}
\value{
a new xts object with the weekly min, max, mean, and sd of the
  adjusted close price for inputted xts_obj
}
\description{
uses endpoints() to split xts_obj into weekly timeseries and then, per weekly
timeseries, find the min, max, mean, sd, of the week price data. Returns a
new xts object with ticker.max, ticker.min, ticker.mean, and ticker.sd as the
columns.
}
