% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_quantmod.R
\name{create_weighted_portfolio}
\alias{create_weighted_portfolio}
\title{create_weighted_portfolio}
\usage{
create_weighted_portfolio(portfolio_returns, weights_vector)
}
\arguments{
\item{weights_vector}{}
}
\description{
Incomplete function: not yet ready to be used.
}
\details{
GM 2/16/2018
}
