% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_quantmod.R
\name{create_xts_transform_of_price_type}
\alias{create_xts_transform_of_price_type}
\title{create_xts_transform_of_price_type}
\usage{
create_xts_transform_of_price_type(xts_obj, time_window, FUN_price_type,
  FUN_transform = mean)
}
\arguments{
\item{xts_obj}{xts object containing time series data to be transformed.}

\item{time_window}{over what time window to transform the data. Usually the
data will be daily so transforms to higher periodicity (e.g. minutes) is
not possible.}

\item{FUN_price_type}{function to pull a specifc price data from the xts_obj:
Ad(), Cl(), Op(), Hi(), Lo()}

\item{FUN_transform}{how to transform the data, what function to use, e.g.
mean, min, max, sd, etc...}
}
\value{
a new xts object
}
\description{
Generic form of create_weekly_max_min_mean_std_of_Ad_price() function. Time
window is generalized, options are: us, secs, mins, minutes, seconds, hours,
days, weeks, months, quarters, years. FUN_transform options are: max, min,
mean, sd, or any other xts transform function, defaulted to "mean". Function
will use endpoints() to split the data by time_window, and then will apply
the transform, and then will merge the xts back and return said xts.
}
