% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_quantmod.R
\name{create_year_month_return_matrix}
\alias{create_year_month_return_matrix}
\title{create_year_month_return_matrix}
\usage{
create_year_month_return_matrix(ticker)
}
\arguments{
\item{ticker}{ticker string of the xts object}
}
\description{
Incomplete function: not yet ready to be used
GM 2/16/2018
}
