% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_quantmod.R
\name{cum_MinMax_AllPrices}
\alias{cum_MinMax_AllPrices}
\title{cum_MinMax_AllPrices}
\usage{
cum_MinMax_AllPrices(xts_obj, period_type = "weeks", num_periods = 1)
}
\arguments{
\item{period_type}{"weeks", "months", "days"}

\item{num_periods}{the number of periods to calculate min and max over.}
}
\value{
a new xts object, with the cummulativce minimum and cummulative
  maximum for the period specified by parameters. The returned xts contains a
  data point for each datapoint in the original input xts. For example, if
  num periods was 4 and period_type was "weeks", then each point in the
  original time series will have a new point that is the minimum and maximum
  of price so far, with the column headers being
  ticker.price_type."cummax"/"cummin"
}
\description{
Takes an xts object of multiple coumns (AD, CL, HI, Lo, OP), or of just one
column/price type, and returns a merged xts object with the cummulative max/min of each
column for the defined period. Uses the built-in functions in xts of cummin
and cummax. Splits the original xts object into period_types of num_periods
(e.g. 1 week = num_periods 1, period_type="weeks")
}
