% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure_api_dataAccess.R
\name{eod_batch_av_helper}
\alias{eod_batch_av_helper}
\title{eod_batch_av_helper}
\usage{
eod_batch_av_helper(ticker, path_to_ticker_dir)
}
\arguments{
\item{path_to_ticker_dir}{}
}
\value{
nothing
}
\description{
Appends new intraday data to data csv files on local respository. Downloads
new timeseries and appends. If a file does not exist on local drive,
downloads full outputsize and writes that out to file.
}
\details{
HARDCODES parameters!!!
}
