% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure_api_dataAccess.R
\name{eod_batch_av_intraday}
\alias{eod_batch_av_intraday}
\title{does a batch download at EOD for the days intraday timeseries. Checks to see
if we have a file for the ticker already on disk. See helper method for more
detail. If list of tickes is NA, then it gets set to the stocks 1000 space,
IWB holdings. List of tickers should be a vector of tickers, and get be
passed in by a call to tq_index or a call to getting a list of tickes from a
csv file.}
\usage{
eod_batch_av_intraday(path_to_ticker_dir, path_to_api_key_file,
  list_of_tickers = NA)
}
\arguments{
\item{list_of_tickers}{}
}
\value{
nothing
}
\description{
Currently, I am using this to do EOD batch runs at night of intraday
timeseries for all ETF's in my list and all IWB holdings.
}
\details{
HARDCODES params!!!
}
