% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_tidy.R
\name{filter_benchmark_by_sector}
\alias{filter_benchmark_by_sector}
\title{filter_benchmark_by_sector}
\usage{
filter_benchmark_by_sector(tibble_obj, sector_name)
}
\arguments{
\item{tibble_obj}{a tibble, must have a column titled sector for this to work.}

\item{sector_name}{the sector name to filter for , e.g. Industrials, Financials, etc...}
}
\value{
a new tibble object with only the sector specified in the sector column.
}
\description{
takes an inputted tibble object and returns a new tibble with the filter
applied. In this function, the filter is sector name, thereby we are
returning a tibble with only the specified sector name in it. The inputted
tibble_obj must have a column titled "sector".
}
