% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_tidy.R
\name{filter_benchmark_by_symbol}
\alias{filter_benchmark_by_symbol}
\title{filter_benchmark_by_symbol}
\usage{
filter_benchmark_by_symbol(tibble_obj, symbol_value)
}
\arguments{
\item{symbol_value}{}
}
\value{
a new tibble object with only the specified symbol in the symbol
  column.
}
\description{
filter the inputted tibble object by symbol. Return only the symbols equal to
the symbol specifid by symbol_value
}
