% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_tidy.R
\name{filter_benchmark_by_weight}
\alias{filter_benchmark_by_weight}
\title{filter_benchmark_by_weight}
\usage{
filter_benchmark_by_weight(tibble_obj, weight_value, op)
}
\arguments{
\item{tibble_obj}{a tibble object, most simply the result of a call to tq_index(<some_index>)}

\item{weight_value}{the value of weight to compare the weight column against.}

\item{op}{options are "==", ">", "<", "<=", etc...}
}
\value{
a tibble with the filtering based on weight applied.
}
\description{
Filter the inputted tibble object by its weighting in the index. The weight
column is returned with retrieving a tibble from tq_index(). The op operator
can be specified, such that we can look for weight values that greater, less
than, or equal to some value, weight_value.
}
