% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_periodicity_to_periodicity.R
\name{from_periodicity_to_periodicity}
\alias{from_periodicity_to_periodicity}
\title{from_periodicity_to_periodicity}
\usage{
from_periodicity_to_periodicity(xts_obj, to_periodicity = "weekly",
  xts_obj_name = NA)
}
\arguments{
\item{to_periodicity}{default to "weekly", could be "monthly", "quarterly",
"yearly", etc...}

\item{xts_obj_name}{default to NA, in which case xts_obj name is used,
usually this would be the ticker}
}
\description{
Convert periodicity of OHLC data. xts_obj is most simply, a result of a call
to getSymbols(). Assumes the column headers are named as getSymbols() would
name them, <ticker>.Open, <ticker>.Close, <ticker>.High, <ticker>.Low, etc...
This function generalizes the xts functions to.<some periodicity>.
}
