% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_quantmod.R
\name{get_adjusted_close_column}
\alias{get_adjusted_close_column}
\title{get_adjusted_close_column}
\usage{
get_adjusted_close_column(stock_xts_obj)
}
\arguments{
\item{stock_xts_obj}{}
}
\value{
the adjusted close column only of the inputted timeseries data
}
\description{
Wrapper for the xts built-in function: Ad()
}
