% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dividends.R
\name{get_all_dividends_for_this_tibble}
\alias{get_all_dividends_for_this_tibble}
\title{get_all_dividends_for_this_tibble}
\usage{
get_all_dividends_for_this_tibble(tibble_obj, index_dir, index_name = "DOW",
  force_pull = FALSE)
}
\arguments{
\item{tibble_obj}{a result of a call to tq_index, most simply}

\item{index_dir}{directory for specified index data, where files such as
historical dividends file reside.}

\item{index_name}{the name of the index for the constituents in the
tibble_obj. See tq_index_options() for possibilities.}

\item{force_pull}{if FALSE, only pulls from remote repository if lcoal file
is older than 1 day, when TRUE, always pulls from remote repository.}
}
\value{
tibble object with dividends and correspoinding dates paid out and
  the symbol of the dividend payer.
}
\description{
Returns a tibble with columns: date, dividends, symbol. Basically a list of
all the dividends paid out for each distinch symbol in the tibble_obj passed
in as parameter. Most simply, the inputted tibble_obj will be a result of a
call to tq_index(). The input tibble must have a column named symbol. Stores
a copy of the dividend tibble in a csv file named <index_name>.dividends.csv.
}
