% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splits.R
\name{get_all_splits_for_this_tibble}
\alias{get_all_splits_for_this_tibble}
\title{get_all_splits_for_this_tibble}
\usage{
get_all_splits_for_this_tibble(tibble_obj)
}
\arguments{
\item{tibble_obj}{}
}
\value{
tibble object containing dates, size of split on said date, and the
  symbol of ticker that split.
}
\description{
Same as get_all_dividends_for_this_tibble except in this case returns splits
for the symbols in the inputted tibble object.
}
