% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_tidy.R
\name{get_all_stocks_in_index}
\alias{get_all_stocks_in_index}
\title{get_all_stocks_in_index}
\usage{
get_all_stocks_in_index(index_name, start_date = "2007-01-01",
  stuff_to_get = c("stock.prices"))
}
\arguments{
\item{index_name}{name of the index to query compoenents, options for this
value are in tq_index_options()}

\item{start_date}{the starting date of the time series data to pull for each
index component}

\item{stuff_to_get}{defaults to stock.prices, but valid options are:
financials, key.ratios, key.stats, dividends, splits, or any cobination of
the latter in a vector format.}
}
\value{
a tibble object containing the time series price data of each stock
  in the specified index.
}
\description{
this is a tidyquant implementation. Get all the components of an index and
pull stuff_to_get. stuff_to_get is defaulted to stock.prices, meaning the
component prices in a tibble time series.
}
