% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure_api_dataAccess.R
\name{get_api_key}
\alias{get_api_key}
\title{get_api_key}
\usage{
get_api_key(api_key_file, vendor = "ALPHAVANTAGE")
}
\arguments{
\item{api_key_file}{path to local file where api keys are stored. Must have
columns labeled "VENDOR" and "API_KEY". Should be a csv file.}

\item{vendor}{options currently are "ALPHAVANTAGE", and "QUANDL".}
}
\value{
an api key as character vector for the specified vendor.
}
\description{
Returns the api key for accessing remote data vendors.
}
