% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R
\name{get_append_daily_log_returns}
\alias{get_append_daily_log_returns}
\title{get_append_daily_log_returns}
\usage{
get_append_daily_log_returns(tibble_obj)
}
\arguments{
\item{tibble_obj}{tibble object, most simply a result of a call to tq_get()}
}
\value{
a new tibble object with the DalyLog.return column appended to the
  original input tibble
}
\description{
Calculates daily returns of price time series. utilizes tq_mutate such that
input tibble object is appended to with a new column, DailyLog.return. This
resulting tibble is then returned.
}
