% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_append_min_max_OHLCVA.R
\name{get_append_min_max_OHLCVA}
\alias{get_append_min_max_OHLCVA}
\title{get_append_min_max_OHLCV}
\usage{
get_append_min_max_OHLCVA(tibble_obj, periodicity = "daily")
}
\arguments{
\item{periodicity}{options are: daily, weekly, monthly, quarterly, yearly}
}
\value{
at tibble with the min and max of each column in the input tibble object
}
\description{
This function will return the min and max of each of the columns in the
inputted tibble for the specified period.
}
