% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R, R/zscore_stats_funcs.R
\name{get_append_monthly_log_returns}
\alias{get_append_monthly_log_returns}
\alias{get_append_monthly_log_returns}
\title{get_append_monthly_log_returns}
\usage{
get_append_monthly_log_returns(tibble_obj, fill_in_between_dates = TRUE)

get_append_monthly_log_returns(tibble_obj, fill_in_between_dates = TRUE)
}
\arguments{
\item{tibble_obj}{most simply, a tibble from a call to tq_get()}

\item{fill_in_between_dates}{TRUE (default), FALSE, no filling of monthly
returns for in-between dates.}

\item{tibble_obj}{most simply, a tibble from a call to tq_get()}

\item{fill_in_between_dates}{TRUE (default), FALSE, no filling of monthly
returns for in-between dates.}
}
\value{
tibble with monthly returns, in addtion to the columns from the input tibble.

tibble with monthly returns, in addtion to the columns from the input tibble.
}
\description{
Calculates montly log returns and appends them to the inputted tibble and
returns this tibble. The returns are returns of adjusted close prices month
over month. When fill_in_between_dates is set to TRUE (default), the returned
tibble will have the original periodicity with the in between dates set to
the last calculated monthly return

Calculates montly log returns and appends them to the inputted tibble and
returns this tibble. The returns are returns of adjusted close prices month
over month. When fill_in_between_dates is set to TRUE (default), the returned
tibble will have the original periodicity with the in between dates set to
the last calculated monthly return
}
