% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R, R/zscore_stats_funcs.R
\name{get_append_quarterly_log_returns}
\alias{get_append_quarterly_log_returns}
\alias{get_append_quarterly_log_returns}
\title{get_append_quarterly_log_returns}
\usage{
get_append_quarterly_log_returns(tibble_obj, fill_in_between_dates = TRUE)

get_append_quarterly_log_returns(tibble_obj, fill_in_between_dates = TRUE)
}
\arguments{
\item{fill_in_between_dates}{TRUE (default) FALSE (to only return quarterly endpoints)}

\item{fill_in_between_dates}{TRUE (default) FALSE (to only return quarterly endpoints)}
}
\value{
tibbler with quarterly returns

tibbler with quarterly returns
}
\description{
Calculates quarterly log returns of adjusted close prices using periodReturns
function of xts. Appends new column of returns to existing tibble and returns
said tibble. If fill_in_between_dates is TRUE (default), the in between dates
will be filled in with the last calculated quarterly return.

Calculates quarterly log returns of adjusted close prices using periodReturns
function of xts. Appends new column of returns to existing tibble and returns
said tibble. If fill_in_between_dates is TRUE (default), the in between dates
will be filled in with the last calculated quarterly return.
}
