% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R, R/zscore_stats_funcs.R
\name{get_append_weekly_log_returns}
\alias{get_append_weekly_log_returns}
\alias{get_append_weekly_log_returns}
\title{get_append_weekly_log_returns}
\usage{
get_append_weekly_log_returns(tibble_obj, fill_in_between_dates = TRUE)

get_append_weekly_log_returns(tibble_obj, fill_in_between_dates = TRUE)
}
\arguments{
\item{tibble_obj}{simply a call to tq_get() will achieve this input}

\item{fill_in_between_dates}{TRUE default}

\item{tibble_obj}{simply a call to tq_get() will achieve this input}

\item{fill_in_between_dates}{TRUE default}
}
\value{
a tibble object with the appended weekly log returns.

a tibble object with the appended weekly log returns.
}
\description{
Same functionality as the get_weekly_log_returns() function but returns the
original tibble_obj with the weekly log returns calculated as a new column.
In addition, it fills in the in-between days with the last weekly log return
that was calculated. This can be turned off by setting the
"fill_in_between_dates" argument to FALSE

Same functionality as the get_weekly_log_returns() function but returns the
original tibble_obj with the weekly log returns calculated as a new column.
In addition, it fills in the in-between days with the last weekly log return
that was calculated. This can be turned off by setting the
"fill_in_between_dates" argument to FALSE
}
