% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R, R/zscore_stats_funcs.R
\name{get_append_yearly_log_returns}
\alias{get_append_yearly_log_returns}
\alias{get_append_yearly_log_returns}
\title{get_append_yearly_log_returns}
\usage{
get_append_yearly_log_returns(tibble_obj, fill_in_between_dates = TRUE)

get_append_yearly_log_returns(tibble_obj, fill_in_between_dates = TRUE)
}
\arguments{
\item{fill_in_between_dates}{}

\item{fill_in_between_dates}{}
}
\description{
Same as get_append_quarterly_log_returns, except yearly

Same as get_append_quarterly_log_returns, except yearly
}
