% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_returns.R
\name{get_benchmark_period_returns}
\alias{get_benchmark_period_returns}
\title{get_benchmark_period_returns}
\usage{
get_benchmark_period_returns(etf_to_sector_mapping_tibble,
  yahoo_stock_prices_dir, start_date = "2017-11-01", periodicity = "daily")
}
\arguments{
\item{start_date}{in string format of: YYYY-MM-DD}

\item{periodicity}{options are: daily, weekly, monthly, quarterly, yearly}
}
\value{
an xts object with the benchmark etf returns.
}
\description{
Returns the period returns for the benchmark tickers listed in the
etf_to_sector_mapping_tibble. We are using quantmod functionality here, not
tidyquant. Returned from this function is an xts object, not a tibble object.
Ideally, we would want to be able to specify source and not hardcode yahoo as
source. Stock price time series is downloaded using quantmod functionality
(getSymbols) where the wrapper we are using here is
"yahoo_Main_retrieve_and_write()" function.
}
