% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_returns_tidy.R
\name{get_benchmark_period_returns_tidy}
\alias{get_benchmark_period_returns_tidy}
\title{get_benchmark_period_returns_tidy}
\usage{
get_benchmark_period_returns_tidy(etf_to_sector_mapping_tibble,
  start_date = "2017-11-01", periodicity = "daily",
  price_type = "adjusted")
}
\arguments{
\item{periodicity}{options are: daily, weekly, monthly, quarterly, yearly}

\item{price_type}{options are: open, high, low, close, adjusted}
}
\value{
benchmark returns in tibble obj
}
\description{
Excactly the same functionality as get_benchmark_period_returns() but instead
of using xts functionality and returning XTS data structure, it uses the
tidyquant functionality and tibble data structure.
}
