% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_quantmod.R
\name{get_close_column}
\alias{get_close_column}
\title{get_close_column}
\usage{
get_close_column(stock_xts_obj)
}
\arguments{
\item{stock_xts_obj}{the xts object of timeseries data}
}
\value{
returns the close column only of inputted time series data.
}
\description{
Wrapper for the xts built-in fucntion: Cl()
}
