% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore_stats_funcs.R
\name{get_close_less_open_div_range}
\alias{get_close_less_open_div_range}
\title{get_close_less_open_div_range}
\usage{
get_close_less_open_div_range(tibble_obj)
}
\arguments{
\item{tibble_obj}{}
}
\value{
tibble with input tibble columns plus a new column denoted "Day.Sharpe".
}
\description{
Return a tibble object, with a new column, "Day.Sharpe". This value is
returned by the helper method close_less_open_div_range(), which is called
inside of a call to tq_mutate().
}
