% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure_api_dataAccess.R
\name{get_daily_adjusted_data_alphavantager}
\alias{get_daily_adjusted_data_alphavantager}
\title{get_daily_adjusted_data_alphavantager}
\usage{
get_daily_adjusted_data_alphavantager(ticker)
}
\arguments{
\item{ticker}{ticker as string}
}
\value{
tibble object with the adjusted close time series data in tibble format.
}
\description{
Get daily adjusted close data from alphavantage api. Returned is a tibble object.
}
