% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure_api_dataAccess.R
\name{get_daily_adjusted_data_av}
\alias{get_daily_adjusted_data_av}
\title{get_daily_adjusted_data_av}
\usage{
get_daily_adjusted_data_av(ticker, av_api_key, start_date)
}
\arguments{
\item{ticker}{ticker as string}

\item{av_api_key}{alphavantage api key, Required.}

\item{start_date}{date to start pulling data, in format YYYY-MM-DD}
}
\value{
xts object with daily periodicity data, starting from start_date
}
\description{
Using the quantmod api, pull daily adjusted close data from alphavantage
repository. Returns an xts object.
}
