% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R
\name{get_daily_log_returns}
\alias{get_daily_log_returns}
\title{get_daily_log_returns}
\usage{
get_daily_log_returns(tibble_obj)
}
\arguments{
\item{tibble_obj}{}
}
\value{
tibble oject containing the daily returns, grouped by symbol.
  Resulting colun name is DailyLog.return.
}
\description{
Takes a input a tibble oject (returned from a call to tq_get()) and returns a
tibble object containing the daily returns, grouped by symbol (ticker). Daily
returns are calculated using the adjusted closing prices, and are geometric
returns (type = log of the periodReturn mutate function). This function
returns a new tibble object with the resulting columns (Daily Returns). It
utilizes tq_transmute vs. tq_mutate.
}
