% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure_api_dataAccess.R
\name{get_files_in_dir}
\alias{get_files_in_dir}
\title{get_files_in_dir}
\usage{
get_files_in_dir(path_to_ticker_dir, pattern = "*.csv")
}
\arguments{
\item{pattern}{a pattern to match, e.g "*.csv".}
}
\value{
the list of files matching the patter in the directory,
  path_to_ticker_dir
}
\description{
Simply a wrapper for list.files. Defaults to get all the csv files in a the
directory, path_to_ticker_dir.
}
