% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_quantmod.R
\name{get_high_column}
\alias{get_high_column}
\title{get_high_column}
\usage{
get_high_column(stock_xts_obj)
}
\arguments{
\item{stock_xts_obj}{the xts object of timeseries data}
}
\value{
the high column only of the inputted time series data.
}
\description{
Wrapper for the xts built-in function: Hi()
}
