% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_data_access.R
\name{get_holdings_from_file_in_tibble_fmt}
\alias{get_holdings_from_file_in_tibble_fmt}
\title{get_holdings_from_file_in_tibble_fmt}
\usage{
get_holdings_from_file_in_tibble_fmt(file_path, sep = ",", n_rows = -1)
}
\arguments{
\item{file_path}{path to locaiton of holdings file}

\item{sep}{default ","}

\item{n_rows}{default read all rows in file}
}
\value{
a tibble file containing the data in the local holdings csv file.
}
\description{
get all columns and rows in tibble format for the local holding file input
}
