% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure_api_dataAccess.R
\name{get_intraday_data_alphavantager}
\alias{get_intraday_data_alphavantager}
\title{get_intraday_data_alphavantager}
\usage{
get_intraday_data_alphavantager(ticker, interval = "1min",
  outputsize = "full", error_cnt = 0)
}
\arguments{
\item{ticker}{ticker as string}

\item{interval}{options are 1min, 5min, 15min, 30min, 60min}
}
\value{
a tibble object containing the alphavantage tick pull data, as
  formatted by alphavantage. See the documentation at alphavantage.co for
  more info.
}
\description{
Calls the av_get method of the alphavantage api and specifies the intraday
time series, where the options for the interval are listed below. This
function is part of the tibble functionality and does not use xts, but the
returned object can be coerced to an xts object.
}
