% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_quantmod.R
\name{get_low_column}
\alias{get_low_column}
\title{get_low_column}
\usage{
get_low_column(stock_xts_obj)
}
\arguments{
\item{stock_xts_obj}{the xts object of timeseries data}
}
\value{
the low column only of the inputted timeseries data
}
\description{
Wrapper for the xts built-in function: Lo()
}
