% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R
\name{get_monthly_log_returns}
\alias{get_monthly_log_returns}
\title{get_monthly_log_returns}
\usage{
get_monthly_log_returns(tibble_obj)
}
\arguments{
\item{tibble_obj}{}
}
\value{
tibble with monthly returns of adjusted close prices
}
\description{
calculates monthly log returns of adjusteed close prices and returns them in
a new tibble.
}
