% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_quantmod.R
\name{get_open_column}
\alias{get_open_column}
\title{get_open_column}
\usage{
get_open_column(stock_xts_obj)
}
\arguments{
\item{stock_xts_obj}{the xts object of timeseries data}
}
\value{
the open column only of the inputted time series data.
}
\description{
Wrapper for the xts built-in function: Op()
}
