% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R
\name{get_period_close_to_next_period_open_return}
\alias{get_period_close_to_next_period_open_return}
\title{get_period_close_to_next_period_open_return}
\usage{
get_period_close_to_next_period_open_return(period_OHLCVA_tibble,
  periodicity = "daily")
}
\arguments{
\item{period_OHLCVA_tibble}{a tibble with the period OHLCVA. Most simply, get
this tibble with a call to get_period_OHLCVA().}

\item{periodicity}{options are : daily, weekly, monthly, quarterly, yearly}
}
\value{
a new tibble object with a column named cl.nxtOp.Ret contained the
  close to next open period returns
}
\description{
Calculates the returns from the close @ period p to the open @ period p+1.
Essentially, calculates O/N returns with the periodicity is daily
}
