% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharpe.R
\name{get_period_close_to_next_period_open_sharpe}
\alias{get_period_close_to_next_period_open_sharpe}
\title{get_period_close_to_next_period_open_sharpe}
\usage{

  get_period_close_to_next_period_open_sharpe(period_close_to_next_period_open_returns_tibble,
  window_size = 5)
}
\arguments{
\item{window_size}{}
}
\description{
get_period_close_to_next_period_open_sharpe
}
