% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharpe.R
\name{get_period_high_to_next_period_low_rolling_sharpe}
\alias{get_period_high_to_next_period_low_rolling_sharpe}
\alias{get_period_high_to_next_period_low_rolling_sharpe}
\title{get_period_high_to_next_period_low_sharpe}
\usage{

  get_period_high_to_next_period_low_rolling_sharpe(period_high_to_next_period_low_returns_tibble,
  window_size = 5)


  get_period_high_to_next_period_low_rolling_sharpe(period_high_to_next_period_low_returns_tibble,
  window_size = 5)
}
\arguments{
\item{window_size}{}

\item{window_size}{}
}
\description{
get_period_high_to_next_period_low_sharpe

get_period_high_to_next_period_low_sharpe
}
