% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R
\name{get_period_low_to_next_period_high_return}
\alias{get_period_low_to_next_period_high_return}
\title{get_period_low_to_next_period_high_return}
\usage{
get_period_low_to_next_period_high_return(period_OHLCVA_tibble,
  periodicity = "daily")
}
\arguments{
\item{period_OHLCVA_tibble}{a tibble containing the OHLCVA for the specified
periodicity, most simply, get this from a call to get_period_OHLCA()}

\item{periodicity}{options are: daily, weekly, monthly, quarterly, yearly.}
}
\value{
a tibble object containing period p low to period p+1 high.
}
\description{
Calculate the returns between the low @ period p to the high @ period p+1.
The input parameter min_max_ohlc_tibble is simply a tibble object returned
from get_append_hi_to_lo_period_returns(). The periodicity inputted is daily,
weekly, monthly, yearly and shoud match that passed to
get_append_hi_to_lo_period_returns(). The reason why we are using the
min_max_ohlc_tibble here as opposed to the original price tibble is because
changing periodicity of a timeseries does not give us the period's high/low,
it gives us the high/low of the endpoint of the period.
}
