% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharpe.R
\name{get_period_low_to_next_period_high_sharpe}
\alias{get_period_low_to_next_period_high_sharpe}
\title{get_period_low_to_next_period_high_sharpe}
\usage{

  get_period_low_to_next_period_high_sharpe(period_low_to_next_period_high_returns_tibble,
  window_size = 5)
}
\arguments{
\item{window_size}{}
}
\description{
get_period_low_to_next_period_high_sharpe
}
