% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_period_OHLCVA.R
\name{get_period_OHLCVA}
\alias{get_period_OHLCVA}
\title{get_period_OHLCVA}
\usage{
get_period_OHLCVA(tibble_obj, periodicity = "daily")
}
\arguments{
\item{tibble_obj}{a result of a call to tq_get(), daily data, most simply}

\item{periodicity}{options are: daily, weekly, monthly, quarterly, yearly}
}
\value{
a new tibble object with OHLCVA rows for the specified periodicity.
}
\description{
Return the open, high, low, close, cum volume, and adjusted close of the
period. This takes the original daily data and returns the first open in the
period, the last close/adj of the period and the max and the min respectively
of the high and low columns for the period.
}
