% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R
\name{get_period_open_to_close_return}
\alias{get_period_open_to_close_return}
\title{get_period_open_to_close_return}
\usage{
get_period_open_to_close_return(period_OHLCVA_tibble, periodicity = "daily")
}
\arguments{
\item{period_OHLCVA_tibble}{a tibble containing OHLCVA data for specified
period, most simply, can be yielded by a call to get_period_OHLCVA()}

\item{periodicity}{options are: daily, weekly, monthly, quarterly, yearly}
}
\value{
a new tibble object with a column for the open_to_close, by period, return.
}
\description{
Accepts as input a tibble of the specified periodicity, and then calculates
the return from the open at period p to the close at period p. This is a same
period return, open to close of the same day/week/month/period.
}
