% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharpe.R
\name{get_period_open_to_close_sharpe}
\alias{get_period_open_to_close_sharpe}
\title{get_period_open_to_close_sharpe}
\usage{
get_period_open_to_close_sharpe(op2cl_returns_tibble, window_size = 5,
  annualize = FALSE, scale = 252)
}
\arguments{
\item{annualize}{defaults to FALSE, if annualize sharpe desird, set to TRUE}

\item{scale}{252 for daily periodicity, 52 for weekly, 12 for monthly.}
}
\value{
tibble object with period open to close sharpe.
}
\description{
get_period_open_to_close_sharpe
}
