% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R
\name{get_quarterly_log_returns}
\alias{get_quarterly_log_returns}
\title{get_quarterly_log_returns}
\usage{
get_quarterly_log_returns(tibble_obj)
}
\arguments{
\item{tibble_obj}{most simply, returned from a call to tq_get()}
}
\value{
a tibble with the quarterly log returns
}
\description{
Calculates the quarterly log returns using xts build in periodReturns
functionality. Returns are quarter over quarter, using adjusted close prices.
}
