% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore_stats_funcs.R
\name{get_returns_helper_function}
\alias{get_returns_helper_function}
\title{get_returns_helper_function}
\usage{
get_returns_helper_function(x, na.rm = TRUE)
}
\arguments{
\item{na.rm}{}
}
\value{
a vector of returns.
}
\description{
x is a vector of prices, the function calculates returns of the series x.
}
