% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore_stats_funcs.R
\name{get_returns_ZScore_stats}
\alias{get_returns_ZScore_stats}
\title{get_returns_ZScore_stats}
\usage{
get_returns_ZScore_stats(tibble_obj, periodicity = "daily", window_size = 5,
  annualize = TRUE, scale = 252)
}
\arguments{
\item{tibble_obj}{input tibble, must at least contain a column <periodicity>Log.return}

\item{window_size}{}
}
\value{
input tibble object, plus columns appended for the zscore stats
  (mean, sd, hi95ci, lo95ci, asr)
}
\description{
Get the mean, standard dev, 95% confidence intervals, and annualized sharpe
ratio of log returns. The input tibble_obj must have a column headered
"<periodicity>Log.return". So if the periodicity of the data is daily, then
there must be a column in the input tibble labeled, "DailyLog.return".
}
