% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_of_change.R
\name{get_rolling_rate_of_change}
\alias{get_rolling_rate_of_change}
\title{get_rolling_rate_of_change}
\usage{
get_rolling_rate_of_change(tibble_obj, price_type = "adjusted",
  periodicity = "daily", num_periods = 1, annualize = FALSE)
}
\arguments{
\item{tibble_obj}{simply, a result of a call to tq_index()}
}
\value{
a new tibble with only the "Rolling" 1 Day return.
}
\description{
Uses the ROC function from TTR library. ROC is rate of change, and is
returned as a percentage value.
}
