% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharpe.R
\name{get_rolling_Sharpe_AdCl_to_AdCl}
\alias{get_rolling_Sharpe_AdCl_to_AdCl}
\title{get_rolling_Sharpe_AdCl_to_AdCl}
\usage{
get_rolling_Sharpe_AdCl_to_AdCl(tibble_obj, num_periods,
  periodicity = "daily", annualize = FALSE, scale = 252)
}
\arguments{
\item{num_periods}{size of rolling window for calc of sharpe ratio}

\item{periodicity}{daily, weekly, monthly, quarterly, yearly}

\item{annualize}{defaults to FALSE, set to TRUE if annualzied sharpe desired}

\item{scale}{252 for daily periodicity, 52 for weekly, 12 for monthly, etc...}
}
\value{
a tibble, all columns from tibble obj input and a new column named "Ad2Ad.Sharpe..."
}
\description{
Calculates a rolling sharpe ratio of returns calculated using adjusted close
prices. The returns are calculated day over day, week over week, month over
month, etc..., depending on the value of periodicity. Finally, we use
tq_mutate to calcularte a rolling non-annualized sharpe ratio, where the size
of the rolling window is specified in the parameter num_periods.
}
