% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_data_access.R
\name{get_tickers_from_local_holdings_file}
\alias{get_tickers_from_local_holdings_file}
\title{get_tickers_from_local_holdings_file}
\usage{
get_tickers_from_local_holdings_file(file_path, sep = ",", n_rows = -1)
}
\arguments{
\item{file_path}{path to file to read containing the holdings as string
tickers.}

\item{sep}{value seperation, defaulted to comma since CSV is expected.}

\item{n_rows}{number of rows to read (post header if header is TRUE),
defaulted to all rows in the file.}

\item{header}{TRUE if file contains a header row, FALSE if not. Defaulted to
TRUE}
}
\value{
a vector of chars, where each element is a ticker.
}
\description{
Accepts as input path to a local CSV file. The default is comma delimited,
containing a header, and reading all rows in the file. If there is no header,
set header to FALSE, and if only a certain number of rows are to be read,
specifiy the number of rows in the nrows parameter. Requires that the column
with the tickers must he headered "Ticker".
}
