% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_quantmod.R
\name{get_volume_column}
\alias{get_volume_column}
\title{get_volumn_column}
\usage{
get_volume_column(stock_xts_obj)
}
\arguments{
\item{stock_xts_obj}{}
}
\value{
the volumn column only of the inputted timeseries data
}
\description{
Wrapper for the xts built-ion function: Vo()
}
