% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R
\name{get_weekly_log_returns}
\alias{get_weekly_log_returns}
\title{get_weekly_log_returns}
\usage{
get_weekly_log_returns(tibble_obj)
}
\arguments{
\item{tibble_obj}{tibble object, a result of a call to tq_get()}
}
\value{
a new tibble object with the weekly log returns as a new column.
}
\description{
Returns the weekly log returns of adjusted close prices, week over week. Uses
the built in xts periodReturn function. More specifically, uses adjusted
close prices of the last day of week t and the last day of week t-1, to yield
a return.
}
