% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns.R
\name{get_yearly_log_returns}
\alias{get_yearly_log_returns}
\title{get_yearly_log_returns}
\usage{
get_yearly_log_returns(tibble_obj)
}
\arguments{
\item{tibble_obj}{}
}
\value{
new tibble with yearly log returns of adjusted close price.
}
\description{
Calculate yearly log returns of adjusted close price.
}
