% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_quantmod.R
\name{merge_created_dataframe_with_original_dataframe}
\alias{merge_created_dataframe_with_original_dataframe}
\title{merge_created_dataframe_with_original_dataframe}
\usage{
merge_created_dataframe_with_original_dataframe(orig_dataframe,
  created_dataframe)
}
\arguments{
\item{created_dataframe}{}
}
\value{
new merged dataframe built from the 2 input dataframes.
}
\description{
This function takes two dataframes and will merge them together using all
default pararms for merge function. Will omit any rows with NA in them using
na.omit() functionality.
}
