% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore_stats_funcs.R
\name{sd_stats_returns_helper}
\alias{sd_stats_returns_helper}
\title{sd_stats_return_helper}
\usage{
sd_stats_returns_helper(x, na.rm = TRUE)
}
\arguments{
\item{na.rm}{}
}
\description{
Calculates the mean and standard deviation of x.
}
