% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations_tidy.R
\name{sector_index_corr_calc_tidy}
\alias{sector_index_corr_calc_tidy}
\title{sector_index_corr_calc_tidy}
\usage{
sector_index_corr_calc_tidy(tidy_etf_returns, window)
}
\arguments{
\item{window}{}
}
\value{
tibble object of sector returns and corresponding correlations
  between sector returns and overall market index returns
}
\description{
Exact same as sector_index_corr_calc but using the tidyquant library and
utilizing tibble objects
}
