% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{sector_index_corr_calc_xts}
\alias{sector_index_corr_calc_xts}
\title{sector_index_corr_calc_xts}
\usage{
sector_index_corr_calc_xts(benchmark_etf_daily_returns, window)
}
\arguments{
\item{benchmark_etf_daily_returns}{xts with period returns for benchmark
sector etf's and overall market etf.}

\item{window}{size of window for rolling correlation calculation in number of
periods (e.g 10 days, 2 weeks)}
}
\value{
xts object with columns for sector etf returns, index etf returns,
  and corresponding correlation between each sector and the overall index.
}
\description{
Calculates the sector versus index correlation, such as SPY versus a sector
specific ETF (XLF) or a single stock. This function is part of a workflow. In
order to get the input parameter benchmark_etf_daily_returns, we need to call
get_benchmark_period_returns(). We then call this function to get the
correlation of the sector etf to the entire index.
}
