% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure_api_dataAccess.R
\name{set_alphavantage_api_key}
\alias{set_alphavantage_api_key}
\title{set_alphavantage_api_key}
\usage{
set_alphavantage_api_key(api_key_file)
}
\value{
nothing
}
\description{
calls the av_api_key() method from the alphavantage library to set the key in
order to pull data from alphavantage
}
