% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure_api_dataAccess.R
\name{set_quandl_api_key}
\alias{set_quandl_api_key}
\title{set_quandl_api_key}
\usage{
set_quandl_api_key(api_key_file)
}
\value{
nothing
}
\description{
sets the quandl api key within the quandl library api so that data can be
pulled from the quandl remote repository.
}
