% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure_api_dataAccess.R
\name{subset_av_tibble}
\alias{subset_av_tibble}
\title{subset_av_tibble}
\usage{
subset_av_tibble(cut_off_date_as_char, path_to_ticker_dir, file_name,
  column_headers = c("BarTimeStamp", "Open", "High", "Low", "Close",
  "Volume"))
}
\arguments{
\item{cut_off_date_as_char}{cut-off date}

\item{path_to_ticker_dir}{path to directory location of ticks file}

\item{file_name}{the name of the ticks file, usually <something>.csv}

\item{column_headers}{the names of the column headers for the alphavantage
ticker file, in a vector. Defaults to c("BarTimeStamp", "Open", "High", "Low", "Close")}
}
\value{
subsetted tibble, contains all the data up to cut off date input
  parameter
}
\description{
subset_av_tibble
}
