% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure_api_dataAccess.R
\name{when_was_mkt_open_last}
\alias{when_was_mkt_open_last}
\title{when_was_mkt_open_last}
\usage{
when_was_mkt_open_last()
}
\value{
a POSIXct object, UTC timezone, of when NY equities was last open
  (the last time the market closed)
}
\description{
Returns the last time the New York Equities market was open. Currently only
supports New York equity markets, but will expand in the future. The last
time the market was open is actually the last time the market closed, we
should change the name. Accounts for dst differences between UTC and EST/EDT.
This function is a work in progress, but is functional right now for NY
Equities
}
