% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_intraday_av.R
\name{write_intraday_av}
\alias{write_intraday_av}
\title{write_intraday_av}
\usage{
write_intraday_av(ticker, data_obj, path_to_ticker_dir, column_names = NULL)
}
\arguments{
\item{column_names}{}
}
\value{
nothing
}
\description{
Write to csv file an intraday time series object retrieved from alphavantage.
}
