% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ticker_csv.R
\name{write_ticker_csv}
\alias{write_ticker_csv}
\title{write_ticker_csv}
\usage{
write_ticker_csv(ticker, path_to_ticker_dir, column_names = NULL)
}
\arguments{
\item{ticker}{the ticker name, used to create the ticker csv filename, and
also to retrieve time seriers data object.}

\item{path_to_ticker_dir}{path to the directory where the ticker csv file
will be written.}

\item{column_names}{column names of the csv file to be written.}
}
\value{
currently returns nothing.
}
\description{
writes an xts time series data object to flat file csv format.
}
