% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore_stats_funcs.R
\name{zscore_stats_helper}
\alias{zscore_stats_helper}
\title{zscore_stats_helper}
\usage{
zscore_stats_helper(x, headers, na.rm = TRUE, annualize = TRUE,
  scale = 252)
}
\arguments{
\item{headers}{a list of headers for the returned liist of values (mean, sd,
upper, lower, asr)}

\item{annualize}{default TRUE, annualize sharpe or not (FALSE)}

\item{scale}{252 or daily, periodicity, 52 for weekly, 12 for monthly, 4 for quarterly, etc..}
}
\value{
a dictionary where the values are keyed by the input list, headers
}
\description{
Calculate the mean, standard deviation, 95% ci upper/lower bound, and
annualized sharpe ratio
}
